Newer fluffos version

+++ fluffos/src/lpc.h
@@ -11,7 +11,7 @@ typedef struct {
 union u {
     const char *string;
     long number;
-    float real;
+    double real;
 
     refed_t *refed; /* any of the block below */
 
@@ -36,7 +36,7 @@ union u {
  */
 typedef struct svalue_s {
     short type;
-    short subtype;
+    int subtype;
     union u u;
 } svalue_t;
